Add hfo detection integration test
